Module 2/Chapter10/django-myproject-10/requirements.txt
Django==1.8.1
django-debug-toolbar==1.4
fabric==1.9.0
awesome-slugify
django-mobile
python-memcached
bpython